+++ packages/suite/src/components/suite/images/DeviceIcon.tsx
