+++ packages/suite/src/components/suite/modals/PassphraseSource.tsx
@@ -3,27 +3,28 @@ import styled from 'styled-components';
 import { ConfirmOnDevice } from '@trezor/components';
 import { Modal } from '@suite-components';
 import { Translation } from '@suite-components/Translation';
+import { DeviceConfirmImage } from '@suite-components/images/DeviceConfirmImage';
 import { useActions } from '@suite-hooks';
 import * as discoveryActions from '@wallet-actions/discoveryActions';
-import DeviceConfirmImage from '@suite-components/images/DeviceConfirmImage';
 import type { TrezorDevice } from '@suite-types';
 
 const StyledModal = styled(Modal)`
     width: 360px;
 `;
 
-interface PassphraseOnDeviceProps {
+interface PassphraseSourceProps {
     device: TrezorDevice;
 }
 
 /**
  * Modal used with model T with legacy firmware as result of 'ButtonRequest_PassphraseType' where passphrase source is requested on device
- * @param {PassphraseOnDeviceProps}
+ * @param {PassphraseSourceProps}
  */
-export const PassphraseOnDevice = ({ device }: PassphraseOnDeviceProps) => {
+export const PassphraseSource = ({ device }: PassphraseSourceProps) => {
     const { getDiscoveryAuthConfirmationStatus } = useActions({
         getDiscoveryAuthConfirmationStatus: discoveryActions.getDiscoveryAuthConfirmationStatus,
     });
+
     const authConfirmation = getDiscoveryAuthConfirmationStatus() || device.authConfirm;
 
     return (
@@ -32,19 +33,10 @@ export const PassphraseOnDevice = ({ device }: PassphraseOnDeviceProps) => {
                 <Translation
                     id={
                         authConfirmation
-                            ? 'TR_CONFIRM_EMPTY_HIDDEN_WALLET_ON'
-                            : 'TR_ENTER_PASSPHRASE_ON_DEVICE_LABEL'
+                            ? 'TR_CONFIRM_PASSPHRASE_SOURCE'
+                            : 'TR_SELECT_PASSPHRASE_SOURCE'
                     }
                     values={{ deviceLabel: device.label }}
-                />
-            }
-            description={
-                <Translation
-                    id={
-                        authConfirmation
-                            ? 'TR_THIS_HIDDEN_WALLET_IS_EMPTY_SOURCE'
-                            : 'TR_PASSPHRASE_CASE_SENSITIVE'
-                    }
                 />
             }
             devicePrompt={
@@ -52,10 +44,13 @@ export const PassphraseOnDevice = ({ device }: PassphraseOnDeviceProps) => {
                     <ConfirmOnDevice
                         title={<Translation id="TR_CONFIRM_ON_TREZOR" />}
                         trezorModel={device.features?.major_version === 1 ? 1 : 2}
+                        animated
                     />
                 ) : null
             }
-            data-test="@modal/enter-passphrase-on-device"
+            data-test={
+                authConfirmation ? '@modal/confirm-empty-hidden-wallet' : '@modal/passphrase-source'
+            }
         >
             <DeviceConfirmImage device={device} />
         </StyledModal>
